apply css variables injection right under the public component
