+++ enterprise/frontend/src/embedding-sdk/components/private/AppInitializeController.tsx
@@ -7,7 +7,7 @@ import { useSdkSelector } from "embedding-sdk/store";
 import { getIsInitialized } from "embedding-sdk/store/selectors";
 import type { SDKConfig } from "embedding-sdk/types";
 
-import { SdkContentWrapper } from "./SdkContentWrapper";
+import { SdkGlobalStylesWrapper } from "./SdkGlobalStylesWrapper";
 
 interface AppInitializeControllerProps {
   children: ReactNode;
@@ -25,11 +25,11 @@ export const AppInitializeController = ({
   const isInitialized = useSdkSelector(getIsInitialized);
 
   return (
-    <SdkContentWrapper
+    <SdkGlobalStylesWrapper
       baseUrl={config.metabaseInstanceUrl}
       id={EMBEDDING_SDK_ROOT_ELEMENT_ID}
     >
       {!isInitialized ? <div>{t`Loading…`}</div> : children}
-    </SdkContentWrapper>
+    </SdkGlobalStylesWrapper>
   );
 };
