+++ frontend/src/metabase/styled-components/containers/GlobalStyles/GlobalStyles.tsx
@@ -6,7 +6,7 @@ import { alpha, color, lighten } from "metabase/lib/colors";
 import { getSitePath } from "metabase/lib/dom";
 import { useSelector } from "metabase/lib/redux";
 import { aceEditorStyles } from "metabase/query_builder/components/NativeQueryEditor/NativeQueryEditor.styled";
-import { useThemeSpecificSelectors } from "metabase/styled-components/theme/theme";
+import { useThemeSpecificCssVariables } from "metabase/styled-components/theme/theme";
 import { saveDomImageStyles } from "metabase/visualizations/lib/save-chart-image";
 
 import { getFont, getFontFiles } from "../../selectors";
@@ -15,7 +15,7 @@ export const GlobalStyles = (): JSX.Element => {
   const font = useSelector(getFont);
   const fontFiles = useSelector(getFontFiles);
 
-  const themeSpecificSelectors = useThemeSpecificSelectors();
+  const themeSpecificSelectors = useThemeSpecificCssVariables();
 
   const sitePath = getSitePath();
